requirements.txt
pycodestyle==2.11.0
locust==2.16.1
pyyaml==6.0.1